russian-weather-tg-bot/requirements.txt
python-telegram-bot==21.3
aiohttp==3.10.5
python-dotenv==1.0.1
Pillow>=10.0.0
playwright>=1.40.0
